// client/src/app/models/cv.model.ts
export interface CV {
  id: string;       
  name: string;  
  surname: string;  
  phone: string;    
  eMail: string;    
  eduName: string;    
  eduFaculty: string;    
  eduLevel: string;    
  eduStatus: string;    
  workName: string;    
  workPosition: string;    
  workLoad: string;    
  workTime: string;    
  skilDesc: string;    
  skilType: string;    
  skilAchievement: string; 
  addrCountry: string;    
  addrIndex: string;    
  addrCity: string;    
  addrStreet: string;    
  addrNr: string;
}